Add alias to include icon, add some checking to ensure sensible useage

+++ tools/include-icon/src/commands/removeIcon.ts
@@ -1,12 +1,15 @@
-/* eslint-disable no-console */
-
 import * as fs from 'fs';
 import * as process from 'process';
 import { resolve, isAbsolute } from 'path';
 import { promisify } from 'util';
 
-import { removeFromIndex } from '../utils/indexFile';
+import {
+    isInIndex,
+    removeAliasFromIndex,
+    removeFromIndex,
+} from '../utils/indexFile';
 import { componentMetadata } from '../utils/componentMetadata';
+import { failure, success } from '../utils/finish';
 
 const removeFile = promisify(fs.unlink);
 
@@ -16,13 +19,28 @@ interface RemoveIconOptions {
 }
 
 export const removeIcon = async ({ name, dir }: RemoveIconOptions) => {
-    const directory = isAbsolute(dir) ? dir : resolve(process.cwd(), dir);
-    const metadata = componentMetadata(name);
+    try {
+        const directory = isAbsolute(dir) ? dir : resolve(process.cwd(), dir);
+        const metadata = componentMetadata(name);
 
-    const filePath = resolve(directory, metadata.path);
+        const filePath = resolve(directory, metadata.path);
 
-    await removeFromIndex(directory, metadata);
-    await removeFile(filePath);
+        const [exists, parent] = await isInIndex(directory, name);
 
-    console.log(`removed ${name} from ${filePath}`);
+        if (!exists) {
+            return failure(`Failed to remove ${name}, icon does not exist.`);
+        }
+
+        if (parent) {
+            await removeAliasFromIndex(directory, parent, name);
+            return success(`Removed alias ${name} from icon ${parent}`);
+        }
+
+        await removeFromIndex(directory, metadata);
+        await removeFile(filePath);
+
+        return success(`Removed icon ${name} in ${filePath}`);
+    } catch (error) {
+        return failure(error);
+    }
 };
